// src/infra/database/common/enums/saga.enum.ts
export enum SagaStatus {
  INITIATED = 'INITIATED',
  EXECUTING = 'EXECUTING',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
  COMPENSATING = 'COMPENSATING',
  COMPENSATED = 'COMPENSATED',
  PARTIALLY_COMPENSATED = 'PARTIALLY_COMPENSATED',
  TIMEOUT = 'TIMEOUT',
  ABORTED = 'ABORTED'
}

export enum SagaStepStatus {
  PENDING = 'PENDING',
  EXECUTING = 'EXECUTING',
  COMPLETED = 'COMPLETED',
  FAILED = 'FAILED',
  COMPENSATING = 'COMPENSATING',
  COMPENSATED = 'COMPENSATED',
  SKIPPED = 'SKIPPED',
  TIMEOUT = 'TIMEOUT'
}

export enum SagaStepType {
  VALIDATION = 'VALIDATION',
  LOCK_ACCOUNT = 'LOCK_ACCOUNT',
  DEBIT = 'DEBIT',
  CREDIT = 'CREDIT',
  EXTERNAL_API = 'EXTERNAL_API',
  NOTIFICATION = 'NOTIFICATION',
  SETTLEMENT = 'SETTLEMENT',
  RELEASE_LOCK = 'RELEASE_LOCK'
}